requirements.txt
alembic==1.4.3
attrs==18.2.0
backcall==0.1.0
bleach==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
click-plugins==1.0.4
cligj==0.5.0
colorama==0.4.1
colormath==3.0.0
cycler==0.10.0
Cython==0.29.3
decorator==4.3.0
defusedxml==0.5.0
descartes==1.1.0
entrypoints==0.3
Fiona==1.8.4
Flask==1.1.2
Flask-API==2.0
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
future==0.17.1
GDAL==2.3.3
geographiclib==1.49
geojson==2.4.1
geopandas==0.4.0
geopy==1.18.1
gunicorn==20.0.4
idna==2.8
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lzstring==1.0.4
Mako==1.1.3
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
multiqc==1.2
munch==2.3.2
nbconvert==5.4.0
nbformat==4.4.0
networkx==1.11
notebook==5.7.4
numpy==1.16.1
oauthlib==3.1.0
osmnx==0.9
overpass==0.6.1
overpy==0.4
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
patsy==0.5.1
pickleshare==0.7.5
prometheus-client==0.5.0
prompt-toolkit==2.0.7
psycopg2==2.8.6
Pygments==2.3.1
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.3.0
pyproj==1.9.6
PySocks==1.7.1
python-dateutil==2.7.5
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2018.9
pywinpty==0.5.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.21.0
requests-oauthlib==1.3.0
Rtree==0.8.3
scikit-learn==0.20.2
scipy==1.2.0
seaborn==0.9.0
Send2Trash==1.5.0
Shapely==1.6.4.post1
simplejson==3.16.0
six==1.12.0
spectra==0.0.11
SQLAlchemy==1.3.2
statsmodels==0.12.1
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
tweepy==3.9.0
urllib3==1.24.1
virtualenv==16.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.4.2
xlrd==1.2.0
